backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0
aiofiles==0.24.0